Update getcsv.py
I increased the condition check as a part of the file reading steps since the output CSV file may contain some lines with blank space when the code is executed in the Windows operating system. To deal with the issue, the condition whether the line is empty is important to the program to have no error during being executed.